psg_mvp_backend/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
certifi==2019.3.9
chardet==3.0.4
coloredlogs==10.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
dataclasses==0.6
defusedxml==0.6.0
Django==2.1.7
django-activity-stream==0.8.0
django-allauth==0.39.1
django-annoying==0.10.4
django-appconf==1.0.3
django-cleanup==3.2.0
django-cors-headers==2.5.2
django-extensions==2.1.6
django-extra-fields==2.0.2
django-imagekit==4.0.2
django-phonenumber-field==2.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.9.2
djongo==1.2.32
elasticsearch==7.0.2
elasticsearch-dsl==7.0.0
googlemaps==3.0.2
hanziconv==0.3.2
humanfriendly==4.18
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
more-itertools==7.0.0
numpy==1.16.4
oauthlib==3.0.1
openapi-codec==1.3.2
pandas==0.24.2
phonenumbers==8.10.10
pilkit==2.0
Pillow==5.4.1
pluggy==0.9.0
py==1.8.0
pymongo==3.7.2
pytest==4.4.1
pytest-cov==2.6.1
pytest-django==3.4.8
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.24.1
